Add SSG for album page
