+++ package.json
@@ -31,7 +31,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-intl": "^5.20.6",
-    "spotify-web-sdk": "^0.7.0",
+    "spotify-web-sdk": "^0.7.1",
     "theme-ui": "^0.10.0"
   },
   "devDependencies": {
