--- pages/album/mock.ts
@@ -1,401 +1,0 @@
-export const albumMock = {
-  album_type: 'album',
-  artists: [
-    {
-      external_urls: {
-        spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-      },
-      href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-      id: '2eam0iDomRHGBypaDQLwWI',
-      name: 'Bleachers',
-      type: 'artist',
-      uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-    },
-  ],
-  copyrights: [
-    {
-      text: '(P) 2021 RCA Records, a division of Sony Music Entertainment',
-      type: 'P',
-    },
-  ],
-  external_ids: {
-    upc: '886449225101',
-  },
-  external_urls: {
-    spotify: 'https://open.spotify.com/album/6SPUtbeCQiPGej0t5RBasE',
-  },
-  genres: [],
-  href: 'https://api.spotify.com/v1/albums/6SPUtbeCQiPGej0t5RBasE',
-  id: '6SPUtbeCQiPGej0t5RBasE',
-  images: [
-    {
-      height: 640,
-      url: 'https://i.scdn.co/image/ab67616d0000b27362f705929d4e45f538005c67',
-      width: 640,
-    },
-    {
-      height: 300,
-      url: 'https://i.scdn.co/image/ab67616d00001e0262f705929d4e45f538005c67',
-      width: 300,
-    },
-    {
-      height: 64,
-      url: 'https://i.scdn.co/image/ab67616d0000485162f705929d4e45f538005c67',
-      width: 64,
-    },
-  ],
-  label: 'RCA Records Label',
-  name: 'Take the Sadness Out of Saturday Night',
-  popularity: 25,
-  release_date: '2021-07-30',
-  release_date_precision: 'day',
-  total_tracks: 10,
-  tracks: {
-    href: 'https://api.spotify.com/v1/albums/6SPUtbeCQiPGej0t5RBasE/tracks?offset=0&limit=50&market=BR',
-    items: [
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 179746,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/6v3wihX7hSonQaBVSYTBgW',
-        },
-        href: 'https://api.spotify.com/v1/tracks/6v3wihX7hSonQaBVSYTBgW',
-        id: '6v3wihX7hSonQaBVSYTBgW',
-        is_local: false,
-        is_playable: true,
-        name: '91',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/293defe8bbec47e425d1f58759c39ea6187c8f54?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 1,
-        type: 'track',
-        uri: 'spotify:track:6v3wihX7hSonQaBVSYTBgW',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/3eqjTLE0HfPfh78zjh6TqT',
-            },
-            href: 'https://api.spotify.com/v1/artists/3eqjTLE0HfPfh78zjh6TqT',
-            id: '3eqjTLE0HfPfh78zjh6TqT',
-            name: 'Bruce Springsteen',
-            type: 'artist',
-            uri: 'spotify:artist:3eqjTLE0HfPfh78zjh6TqT',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 244186,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/1AA3ZjLo9tD2iSZAs2svyj',
-        },
-        href: 'https://api.spotify.com/v1/tracks/1AA3ZjLo9tD2iSZAs2svyj',
-        id: '1AA3ZjLo9tD2iSZAs2svyj',
-        is_local: false,
-        is_playable: true,
-        name: 'Chinatown (feat. Bruce Springsteen)',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/ab69115a0809e1e8777f4282c77a5b5eff4cb90f?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 2,
-        type: 'track',
-        uri: 'spotify:track:1AA3ZjLo9tD2iSZAs2svyj',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 240200,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/4lKg94fHZZ3pozsNFTew2x',
-        },
-        href: 'https://api.spotify.com/v1/tracks/4lKg94fHZZ3pozsNFTew2x',
-        id: '4lKg94fHZZ3pozsNFTew2x',
-        is_local: false,
-        is_playable: true,
-        name: 'How Dare You Want More',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/a691ed46cf90a959279bbc769870b4aa94daeeb8?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 3,
-        type: 'track',
-        uri: 'spotify:track:4lKg94fHZZ3pozsNFTew2x',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 150480,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/4M3Uqqw3L6oGEMk69SYIKE',
-        },
-        href: 'https://api.spotify.com/v1/tracks/4M3Uqqw3L6oGEMk69SYIKE',
-        id: '4M3Uqqw3L6oGEMk69SYIKE',
-        is_local: false,
-        is_playable: true,
-        name: 'Big Life',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/f459563c91973db6a9233c16c8a48a280e16b2de?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 4,
-        type: 'track',
-        uri: 'spotify:track:4M3Uqqw3L6oGEMk69SYIKE',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/00FQb4jTyendYWaN8pK0wa',
-            },
-            href: 'https://api.spotify.com/v1/artists/00FQb4jTyendYWaN8pK0wa',
-            id: '00FQb4jTyendYWaN8pK0wa',
-            name: 'Lana Del Rey',
-            type: 'artist',
-            uri: 'spotify:artist:00FQb4jTyendYWaN8pK0wa',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 186613,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/0dagxop90zpA32LDwjAGVk',
-        },
-        href: 'https://api.spotify.com/v1/tracks/0dagxop90zpA32LDwjAGVk',
-        id: '0dagxop90zpA32LDwjAGVk',
-        is_local: false,
-        is_playable: true,
-        name: 'Secret Life (feat. Lana Del Rey)',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/1ab70cfe0fb70eb8efe61f0d346375c3b3199d91?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 5,
-        type: 'track',
-        uri: 'spotify:track:0dagxop90zpA32LDwjAGVk',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 199626,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/7fRCD4vVNpCy91Y3zxNMUl',
-        },
-        href: 'https://api.spotify.com/v1/tracks/7fRCD4vVNpCy91Y3zxNMUl',
-        id: '7fRCD4vVNpCy91Y3zxNMUl',
-        is_local: false,
-        is_playable: true,
-        name: 'Stop Making This Hurt',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/f947c97c1d86bad8ceaeec883fa195fd53b7f65e?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 6,
-        type: 'track',
-        uri: 'spotify:track:7fRCD4vVNpCy91Y3zxNMUl',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 243946,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/6ZpkPahiQA2u63Jx9TnYNw',
-        },
-        href: 'https://api.spotify.com/v1/tracks/6ZpkPahiQA2u63Jx9TnYNw',
-        id: '6ZpkPahiQA2u63Jx9TnYNw',
-        is_local: false,
-        is_playable: true,
-        name: "Don't Go Dark",
-        preview_url:
-          'https://p.scdn.co/mp3-preview/5ae2d1093805683e9e4749ca23fa347fbac4d223?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 7,
-        type: 'track',
-        uri: 'spotify:track:6ZpkPahiQA2u63Jx9TnYNw',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 203533,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/6GXxOycOvD7pp0F4ee6X5R',
-        },
-        href: 'https://api.spotify.com/v1/tracks/6GXxOycOvD7pp0F4ee6X5R',
-        id: '6GXxOycOvD7pp0F4ee6X5R',
-        is_local: false,
-        is_playable: true,
-        name: '45',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/b2eeab7fd7cafe3c942f8260c572c58a2464c156?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 8,
-        type: 'track',
-        uri: 'spotify:track:6GXxOycOvD7pp0F4ee6X5R',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 185653,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/6beGj0daOi7LDVSRm4bPFW',
-        },
-        href: 'https://api.spotify.com/v1/tracks/6beGj0daOi7LDVSRm4bPFW',
-        id: '6beGj0daOi7LDVSRm4bPFW',
-        is_local: false,
-        is_playable: true,
-        name: 'Strange Behavior',
-        preview_url:
-          'https://p.scdn.co/mp3-preview/75e99804ce5fed1bdfcae708b112845ad8baee92?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 9,
-        type: 'track',
-        uri: 'spotify:track:6beGj0daOi7LDVSRm4bPFW',
-      },
-      {
-        artists: [
-          {
-            external_urls: {
-              spotify: 'https://open.spotify.com/artist/2eam0iDomRHGBypaDQLwWI',
-            },
-            href: 'https://api.spotify.com/v1/artists/2eam0iDomRHGBypaDQLwWI',
-            id: '2eam0iDomRHGBypaDQLwWI',
-            name: 'Bleachers',
-            type: 'artist',
-            uri: 'spotify:artist:2eam0iDomRHGBypaDQLwWI',
-          },
-        ],
-        disc_number: 1,
-        duration_ms: 198253,
-        explicit: false,
-        external_urls: {
-          spotify: 'https://open.spotify.com/track/3fSH0czOBg0XYjkviSL6ey',
-        },
-        href: 'https://api.spotify.com/v1/tracks/3fSH0czOBg0XYjkviSL6ey',
-        id: '3fSH0czOBg0XYjkviSL6ey',
-        is_local: false,
-        is_playable: true,
-        name: "What'd I Do With All This Faith?",
-        preview_url:
-          'https://p.scdn.co/mp3-preview/bf0177c319be7b7019c743cf704ee60a4b60eaa4?cid=774b29d4f13844c495f206cafdad9c86',
-        track_number: 10,
-        type: 'track',
-        uri: 'spotify:track:3fSH0czOBg0XYjkviSL6ey',
-      },
-    ],
-    limit: 50,
-    next: null,
-    offset: 0,
-    previous: null,
-    total: 10,
-  },
-  type: 'album',
-  uri: 'spotify:album:6SPUtbeCQiPGej0t5RBasE',
-}
-
-export const reviewMock = {
-  id: 'id',
-  post: `You called me again, drunk in your Benz, driving home under the influence.
-  You scared me to death but I am wasting my breath, because you only listen
-  to your f-ing friends... I do not relate to you, I do not relate to you,
-  no! Because I would never treat me this shitty, you made me hate this
-  city.`,
-  rating: '⭐️⭐️⭐️⭐️⭐️',
-  updatedAt: new Date(),
-  author: {
-    image:
-      'https://pbs.twimg.com/profile_images/1405947624184713220/aCLC4bTE_400x400.jpg',
-    name: 'Tayla Shift',
-  },
-}
