+++ pages/albums/mock.ts
@@ -1,0 +1,15 @@
+export const reviewMock = {
+  id: 'id',
+  post: `You called me again, drunk in your Benz, driving home under the influence.
+  You scared me to death but I am wasting my breath, because you only listen
+  to your f-ing friends... I do not relate to you, I do not relate to you,
+  no! Because I would never treat me this shitty, you made me hate this
+  city.`,
+  rating: '⭐️⭐️⭐️⭐️⭐️',
+  updatedAt: new Date(),
+  author: {
+    image:
+      'https://pbs.twimg.com/profile_images/1405947624184713220/aCLC4bTE_400x400.jpg',
+    name: 'Tayla Shift',
+  },
+}
